Silence annoying warnings with extra cast.

+++ deko/AEMacros.h
@@ -1,4 +1,4 @@
-#define AELOG_LOG(levelName, fmt, ...) NSLog((@"%@ [T:0x%x %@] %s:%d " fmt), levelName, (unsigned int)[NSThread currentThread], ([[NSThread currentThread] isMainThread] ? @"M" : @"S"), __PRETTY_FUNCTION__, __LINE__, ##__VA_ARGS__)
+#define AELOG_LOG(levelName, fmt, ...) NSLog((@"%@ [T:0x%x %@] %s:%d " fmt), levelName, (unsigned int)(unsigned long)[NSThread currentThread], ([[NSThread currentThread] isMainThread] ? @"M" : @"S"), __PRETTY_FUNCTION__, __LINE__, ##__VA_ARGS__)
 
 #define AELOG_ERROR(fmt, ...) AELOG_LOG(@"ERROR", fmt, ##__VA_ARGS__)
